no warranty whatsoever.

+++ src/pong/net/IGameDataHandler.java
@@ -2,5 +2,5 @@ package pong.net;
 
 public interface IGameDataHandler
 {
-    void dataArrived(int[] data);
+    void dataArrived(float[] data);
 }
